Got full generation of libclang c header files to odin, with a quick test to see if it errors (which it didn't)
I'm very excited that I was able to get it to work fully!

I've got typedefs being able to transfer to proper definitions in Odin and I think that's neat.

Of course there are definitions like time_t or __time_64_t, which I can't generate knowingly since it comes from system headers, and I specified the queue to only pull in asts that are NOT system headers. I don't think it will be much of an issue though, it generates most of it correct first time, only with a few errors, which is not bad.